# use multiprocessing in validation and analysis for speedup