Validation running time went from 100s to 20s on my machine (10s for threshold computation and 10s for validation)

Each validation now runs in parallel in a separate pool.
Each simulation in iv_curve and fi_curve also run in parallel in a separate pool